Started working on a bot tests for pnacl
Review URL: http://codereview.chromium.org/2050015

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@2260 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 04204d0e464201bb7725689cfe187a033316ff96

+++ tools/llvm/untrusted-toolchain-creator.sh
@@ -757,7 +757,12 @@ BuildAndInstallBinutils() {
 BuildAndInstallNewlib() {
   local saved_dir=$(pwd)
   local tmpdir=${TMP}/newlib
-  rm -rf ${NEWLIB_INSTALL_DIR}
+  # NOTE: When an argument is given to this fucntion we abuse it to generate
+  #       a bitcode version of the library. In the scenario we just want to
+  #       build libc.a and do install or uninstall anything.
+  if [ $# == 0 ] ; then
+    rm -rf ${NEWLIB_INSTALL_DIR}
+  fi
   Banner "building and installing newlib"
 
   cd ${tmpdir}/newlib-1.17.0
